# client-ts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VISCOTTA Shipping API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/spedizioni/valida/{order_number}": {
      "post": {
        "summary": "Valida Spedizione",
        "description": "Quota la spedizione di un ordine reale via MyDHL API (POST /rates) —\nnessuna spedizione viene creata, nessuna etichetta emessa: è la\n\"bozza\" emulata. Combina cartonizzazione (pesi scatoloni) e indirizzo\ncliente da DB.",
        "operationId": "valida_spedizione_spedizioni_valida__order_number__post",
        "parameters": [
          {
            "name": "order_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Valida Spedizione Spedizioni Valida  Order Number  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/spedizioni": {
      "post": {
        "summary": "Crea Bozza Spedizione",
        "description": "Crea la bozza (stato 'bozza'): quota via /rates e salva su DB.\nNessuna chiamata DHL con effetto reale — solo /rates, come /valida.",
        "operationId": "crea_bozza_spedizione_spedizioni_post",
        "parameters": [
          {
            "name": "order_number",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpedizioneResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/spedizioni/elenco": {
      "get": {
        "summary": "Elenco Spedizioni",
        "description": "Vista d'insieme, sola lettura: tutti gli ordini \"in prenotazione\" nel\nperiodo (default: oggi + 13 giorni, due settimane), con lo stato dei\ncolli scansionati e della spedizione — se esiste già una bozza/una\nconferma. Nessuna azione qui: solo per farsi un'idea di cosa c'è da\nprocessare, prima di agire con le chiamate dedicate.",
        "operationId": "elenco_spedizioni_spedizioni_elenco_get",
        "parameters": [
          {
            "name": "data_da",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Data Da"
            }
          },
          {
            "name": "data_a",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Data A"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RigaElenco"
                  },
                  "title": "Response Elenco Spedizioni Spedizioni Elenco Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/spedizioni/per-ordine/{order_number}": {
      "get": {
        "summary": "Spedizione Per Ordine",
        "description": "Spedizione più recente per un ordine, o null se non è mai stata\ncreata una bozza — non è un errore, è lo stato \"da iniziare\". Usato\ndalla pagina di dettaglio ordine, che non ha altrimenti modo di\nrisalire allo spedizione_id partendo dal solo order_number.",
        "operationId": "spedizione_per_ordine_spedizioni_per_ordine__order_number__get",
        "parameters": [
          {
            "name": "order_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/SpedizioneResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Spedizione Per Ordine Spedizioni Per Ordine  Order Number  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/spedizioni/{spedizione_id}": {
      "get": {
        "summary": "Dettaglio Spedizione",
        "operationId": "dettaglio_spedizione_spedizioni__spedizione_id__get",
        "parameters": [
          {
            "name": "spedizione_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Spedizione Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpedizioneResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Elimina Bozza Spedizione",
        "description": "Cancella SOLO se ancora in stato 'bozza' — nessuna chiamata DHL con\neffetto reale è mai avvenuta per una bozza, quindi non c'è nulla da\nannullare lato DHL.",
        "operationId": "elimina_bozza_spedizione_spedizioni__spedizione_id__delete",
        "parameters": [
          {
            "name": "spedizione_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Spedizione Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/spedizioni/{spedizione_id}/conferma": {
      "post": {
        "summary": "Conferma Spedizione",
        "description": "bozza -> confermata. QUESTA CHIAMATA HA EFFETTO REALE: crea la\nspedizione DHL vera con relativo costo (dhl.crea_spedizione). Va\ninvocata solo dopo controllo umano della bozza (GET /spedizioni/{id}).\n\nGate: rifiuta (409) se il reparto non ha ancora scansionato tutti i\ncolli dell'ordine (vedi POST /cartonizzazioni/colli/conferma) — senza\nquesto controllo si potrebbe confermare (costo reale, ritiro reale)\nuna spedizione con uno scatolone ancora aperto sul tavolo.",
        "operationId": "conferma_spedizione_spedizioni__spedizione_id__conferma_post",
        "parameters": [
          {
            "name": "spedizione_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Spedizione Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpedizioneResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/spedizioni/{spedizione_id}/etichetta": {
      "get": {
        "summary": "Etichetta Spedizione",
        "description": "PDF etichetta ufficiale DHL — solo per spedizioni 'confermata'/'ritirata'.",
        "operationId": "etichetta_spedizione_spedizioni__spedizione_id__etichetta_get",
        "parameters": [
          {
            "name": "spedizione_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Spedizione Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/spedizioni/{spedizione_id}/pickup": {
      "post": {
        "summary": "Richiedi Pickup Spedizione",
        "description": "confermata -> ritirata. QUESTA CHIAMATA HA EFFETTO REALE: prenota\nil ritiro DHL vero (dhl.richiedi_pickup). data_pickup di default\ndomani (i ritiri lo stesso giorno spesso non sono più prenotabili\ndopo il cutoff — vedi pickupCapabilities in /rates).",
        "operationId": "richiedi_pickup_spedizione_spedizioni__spedizione_id__pickup_post",
        "parameters": [
          {
            "name": "spedizione_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Spedizione Id"
            }
          },
          {
            "name": "data_pickup",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Data Pickup"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpedizioneResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RigaElenco": {
        "properties": {
          "order_number": {
            "type": "string",
            "title": "Order Number"
          },
          "cliente": {
            "type": "string",
            "title": "Cliente"
          },
          "data_consegna": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Consegna"
          },
          "n_colli": {
            "type": "integer",
            "title": "N Colli"
          },
          "colli_confermati": {
            "type": "integer",
            "title": "Colli Confermati"
          },
          "colli_completo": {
            "type": "boolean",
            "title": "Colli Completo"
          },
          "spedizione_stato": {
            "type": "string",
            "title": "Spedizione Stato"
          },
          "spedizione_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spedizione Id"
          }
        },
        "type": "object",
        "required": [
          "order_number",
          "cliente",
          "data_consegna",
          "n_colli",
          "colli_confermati",
          "colli_completo",
          "spedizione_stato",
          "spedizione_id"
        ],
        "title": "RigaElenco"
      },
...
      "SpedizioneResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "order_number": {
            "type": "string",
            "title": "Order Number"
          },
          "corriere": {
            "type": "string",
            "title": "Corriere"
          },
          "stato": {
            "type": "string",
            "title": "Stato"
          },
          "product_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Code"
          },
          "pesi_scatoloni_kg": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Pesi Scatoloni Kg"
          },
          "prezzo_stimato_eur": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prezzo Stimato Eur"
          },
          "shipment_tracking_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shipment Tracking Number"
          },
          "tracking_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracking Url"
          },
          "dispatch_confirmation_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dispatch Confirmation Number"
          },
          "errore": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errore"
          },
          "creata_at": {
            "type": "string",
            "title": "Creata At"
          },
          "confermata_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confermata At"
          },
          "ritirata_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ritirata At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "order_number",
          "corriere",
          "stato",
          "product_code",
          "pesi_scatoloni_kg",
          "prezzo_stimato_eur",
          "shipment_tracking_number",
          "tracking_url",
          "dispatch_confirmation_number",
          "errore",
          "creata_at",
          "confermata_at",
          "ritirata_at"
        ],
        "title": "SpedizioneResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}